requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=0.12.0=pypi_0
aiofiles=0.6.0=pypi_0
aiohttp=3.7.4.post0=pypi_0
albumentations=0.5.2=pypi_0
anyio=2.2.0=py38h50d1736_0
appnope=0.1.2=py38h50d1736_1
argon2-cffi=20.1.0=py38h5406a74_2
async-timeout=3.0.1=pypi_0
async_generator=1.10=py_0
attrs=20.3.0=pyhd3deb0d_0
babel=2.9.0=pyhd3deb0d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.1=py_0
blas=1.0=mkl
bleach=3.3.0=pyh44b312d_0
brotlipy=0.7.0=py38h5406a74_1001
bzip2=1.0.8=h1de35cc_0
ca-certificates=2020.12.5=h033912b_0
cachetools=4.2.1=pypi_0
certifi=2020.12.5=py38h50d1736_1
cffi=1.14.5=py38ha97d567_0
chardet=4.0.0=py38h50d1736_1
cryptography=3.4.4=py38h43df06b_0
cycler=0.10.0=pypi_0
decorator=4.4.2=py_0
defusedxml=0.7.0=pyhd8ed1ab_0
efficientnet-pytorch=0.7.0=pypi_0
entrypoints=0.3=pyhd8ed1ab_1003
ffmpeg=4.3=h0a44026_0
freetype=2.10.4=ha233b18_0
fsspec=0.9.0=pypi_0
future=0.18.2=pypi_0
gettext=0.19.8.1=hb0f4f8b_2
gmp=6.2.1=h23ab428_2
gnutls=3.6.5=h91ad68e_1002
google-api-core=1.26.3=pypi_0
google-auth=1.28.1=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
google-cloud-vision=2.3.0=pypi_0
googleapis-common-protos=1.53.0=pypi_0
grpcio=1.37.0=pypi_0
idna=2.10=pyh9f0ad1d_0
imagehash=4.2.0=pypi_0
imageio=2.9.0=pypi_0
imgaug=0.4.0=pypi_0
importlib-metadata=3.7.0=py38h50d1736_0
install=1.3.4=pypi_0
intel-openmp=2019.4=233
ipykernel=5.5.0=py38h9bb44b7_1
ipython=7.21.0=py38h9bb44b7_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.6.3=pyhd3deb0d_0
jedi=0.18.0=py38h50d1736_2
jinja2=2.11.3=pyh44b312d_0
joblib=1.0.1=pypi_0
jpeg=9b=he5867d9_2
json5=0.9.5=pyh9f0ad1d_0
jsonschema=3.2.0=pyhd8ed1ab_3
jupyter-client=6.1.12=pypi_0
jupyter-packaging=0.7.12=pyhd8ed1ab_0
jupyter_core=4.7.1=py38h50d1736_0
jupyter_server=1.4.1=py38h50d1736_0
jupyterlab=3.0.9=pyhd8ed1ab_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_server=2.3.0=pyhd8ed1ab_0
jupyterlab_widgets=1.0.0=pyhd8ed1ab_1
kiwisolver=1.3.1=pypi_0
lame=3.100=h1de35cc_0
lcms2=2.11=h92f6f08_0
libcxx=10.0.0=1
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libiconv=1.16=h1de35cc_0
libpng=1.6.37=ha441bb4_0
libsodium=1.0.18=hbcb3906_1
libtiff=4.1.0=hcb84e12_1
libuv=1.40.0=haf1e3a3_0
lz4-c=1.9.3=h23ab428_0
markdown=3.3.4=pypi_0
markupsafe=1.1.1=py38h5406a74_3
matplotlib=3.3.4=pypi_0
mistune=0.8.4=py38h5406a74_1003
mkl=2019.4=233
mkl-service=2.3.0=py38h9ed2024_0
mkl_fft=1.3.0=py38ha059aab_0
mkl_random=1.1.1=py38h959d312_0
multidict=5.1.0=pypi_0
nbclassic=0.2.6=pyhd8ed1ab_0
nbclient=0.5.3=pyhd8ed1ab_0
nbconvert=6.0.7=py38h50d1736_3
nbformat=5.1.2=pyhd8ed1ab_1
ncurses=6.2=h0a44026_1
nest-asyncio=1.4.3=pyhd8ed1ab_0
nettle=3.4.1=h3018a27_0
networkx=2.5.1=pypi_0
ninja=1.10.2=py38hf7b0b51_0
notebook=6.2.0=py38h50d1736_0
numpy=1.20.1=pypi_0
numpy-base=1.19.2=py38hcfb5961_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py_0
opencv-python=4.5.1.48=pypi_0
opencv-python-headless=4.5.1.48=pypi_0
openh264=2.1.0=hd9629dc_0
openssl=1.1.1k=h0d85af4_0
packaging=20.9=pyh44b312d_0
pandas=1.2.3=pypi_0
pandoc=2.11.4=h35c211d_0
pandocfilters=1.4.2=py_1
parso=0.8.1=pyhd8ed1ab_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.2.0=pypi_0
pip=21.0.1=py38hecd8cb5_0
prometheus_client=0.9.0=pyhd3deb0d_0
prompt-toolkit=3.0.16=pyha770c72_0
proto-plus=1.18.1=pypi_0
protobuf=3.15.8=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.8.0=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyrsistent=0.17.3=py38h5406a74_2
pysocks=1.7.1=py38h50d1736_3
pytesseract=0.3.7=pyhd3deb0d_0
python=3.8.8=h88f2d9e_4
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytorch=1.8.0=py3.8_0
pytorch-lightning=1.2.7=pypi_0
pytz=2021.1=pyhd8ed1ab_0
pywavelets=1.1.1=pypi_0
pyyaml=5.3.1=pypi_0
pyzmq=20.0.0=py38h23ab428_1
readline=8.1=h9ed2024_0
requests=2.25.1=pyhd3deb0d_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
scikit-image=0.18.1=pypi_0
scikit-learn=0.24.1=pypi_0
scipy=1.6.2=pypi_0
send2trash=1.5.0=py_0
setuptools=52.0.0=py38hecd8cb5_0
shapely=1.7.1=pypi_0
six=1.15.0=pyh9f0ad1d_0
sniffio=1.2.0=py38h50d1736_1
sqlite=3.33.0=hffcf06c_0
tensorboard=2.4.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
terminado=0.9.2=py38h50d1736_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pypi_0
tifffile=2021.4.8=pypi_0
tk=8.6.10=hb0a8c7a_0
torchaudio=0.8.0=py38
torchmetrics=0.2.0=pypi_0
torchvision=0.9.0=py38_cpu
tornado=6.1=py38h5406a74_1
tqdm=4.60.0=pypi_0
traitlets=5.0.5=py_0
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.26.3=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py38h50d1736_4
xz=5.2.5=h1de35cc_0
yarl=1.6.3=pypi_0
zeromq=4.3.3=hb1e8313_3
zipp=3.4.1=pyhd8ed1ab_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.5=h41d2c2f_0
